feat: new points fetching method, start on new leaderboard table wip
